+++ src/network/requestItem.ts
@@ -1,4 +1,6 @@
-import * as tool from '../lib/tool';
+
+import { getUniqueID } from '../lib/tool';
+import { genResonseByResponseType, genGetDataByUrl } from './helper';
 
 export type VConsoleRequestMethod = '' | 'GET' | 'POST' | 'PUT' | 'DELETE' | 'HEAD' | 'CONNECT' | 'OPTIONS' | 'TRACE' | 'PATCH';
 
@@ -7,23 +9,27 @@ export class VConsoleNetworkRequestItem {
   name?: string              = '';
   method: VConsoleRequestMethod = '';
   url: string                = '';
-  status: number | string    = 0;
-  statusText?: string        = '';
+  status: number | string    = 0; // HTTP status codes
+  statusText?: string        = ''; // for display
+  cancelState?: 0 | 1 | 2 | 3 = 0; // 0=no cancel; 1=abort (for XHR); 2=cancel (for Fetch); 3=timeout;
   readyState?: XMLHttpRequest['readyState'] = 0;
   header: { [key: string]: string } = null; // response header
   responseType: XMLHttpRequest['responseType'] = '';
   requestType: 'xhr' | 'fetch' | 'ping' | 'custom';
   requestHeader: HeadersInit = null;
   response: any;
+  responseSize: number      = 0; // bytes
+  responseSizeText: string  = '';
   startTime: number         = 0;
   endTime: number           = 0;
   costTime?: number         = 0;
   getData: { [key: string]: string } = null;
   postData: { [key: string]: string } | string = null;
   actived: boolean          = false;
+  noVConsole?: boolean      = false;
 
   constructor() {
-    this.id = tool.getUniqueID();
+    this.id = getUniqueID();
   }
 }
 
@@ -43,7 +49,7 @@ export class VConsoleNetworkRequestItemProxy extends VConsoleNetworkRequestItem 
           // NOTICE: Once the `response` is set, 
           //         modifying its internal properties will not take effect,
           //         unless a new `response` is re-assigned.
-          item._response = RequestItemHelper.genResonseByResponseType(item.responseType, value);
+          item._response = genResonseByResponseType(item.responseType, value);
           return true;
           
         case 'url':
@@ -51,7 +57,7 @@ export class VConsoleNetworkRequestItemProxy extends VConsoleNetworkRequestItem 
           const name = value?.replace(new RegExp('[/]*$'), '').split('/').pop() || 'Unknown';
           Reflect.set(item, 'name', name);
 
-          const getData = RequestItemHelper.genGetDataByUrl(value, item.getData);
+          const getData = genGetDataByUrl(value, item.getData);
           Reflect.set(item, 'getData', getData);
           break;
         case 'status':
@@ -84,68 +90,3 @@ export class VConsoleNetworkRequestItemProxy extends VConsoleNetworkRequestItem 
     return new Proxy(item, VConsoleNetworkRequestItemProxy.Handler);
   }
 }
-
-export const RequestItemHelper = {
-  /**
-   * Generate `getData` by url. 
-   */
-  genGetDataByUrl(url: string, getData = {}) {
-    if (!tool.isObject(getData)) {
-      getData = {};
-    }
-    let query: string[] = url ? url.split('?') : []; // a.php?b=c&d=?e => ['a.php', 'b=c&d=', 'e']
-    query.shift(); // => ['b=c&d=', 'e']
-    if (query.length > 0) {
-      query = query.join('?').split('&'); // => 'b=c&d=?e' => ['b=c', 'd=?e']
-      for (const q of query) {
-        const kv = q.split('=');
-        try {
-          getData[ kv[0] ] = decodeURIComponent(kv[1]);
-        } catch (e) {
-          // "URIError: URI malformed" will be thrown when `kv[1]` contains "%", so just use raw data
-          // @issue #470
-          // @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Malformed_URI
-          getData[ kv[0] ] = kv[1];
-        }
-      }
-    }
-    return getData;
-  },
-
-  /**
-   * Generate formatted response data by responseType.
-   */
-  genResonseByResponseType(responseType: string, response) {
-    let ret;
-    switch (responseType) {
-      case '':
-      case 'text':
-      case 'json':
-        // try to parse JSON
-        if (tool.isString(response)) {
-          try {
-            ret = JSON.parse(response);
-            ret = tool.safeJSONStringify(ret, { maxDepth: 10, keyMaxLen: 500000, pretty: true });
-          } catch (e) {
-            // not a JSON string
-            ret = response;
-          }
-        } else if (tool.isObject(response) || tool.isArray(response)) {
-          ret = tool.safeJSONStringify(response, { maxDepth: 10, keyMaxLen: 500000, pretty: true });
-        } else if (typeof response !== 'undefined') {
-          ret = Object.prototype.toString.call(response);
-        }
-        break;
-  
-      case 'blob':
-      case 'document':
-      case 'arraybuffer':
-      default:
-        if (typeof response !== 'undefined') {
-          ret = Object.prototype.toString.call(response);
-        }
-        break;
-    }
-    return ret;
-  },
-};